add scroll into view while editing functionality
